requirements.txt
fastapi
uvicorn
torch
torchvision
Pillow
numpy
opencv-python-headless
